+++ config.ts
@@ -1,0 +1,9 @@
+const config = {
+  highText: {
+    style: {
+      color: '#FF4D4F'
+    }
+  }
+}
+
+export default config