+++ index.ts
@@ -1,11 +1,3 @@
-import AutoIndex from "./src/AutoIndex.vue"
-import HighText from "./src/HighText.vue"
-import InputFile from "./src/InputFile.vue"
-import ShowValue from "./src/ShowValue.vue"
 
 export {
-  AutoIndex,
-  HighText,
-  InputFile,
-  ShowValue
 }