+++ history.md
@@ -1,4 +1,4 @@
 ### ToDo
 
 ### 4.0.1
-- 基于v4版本的complex-data重构component
+- 重构component