+++ history.md
@@ -1,4 +1,4 @@
 ### ToDo
 
 ### 4.0.1
-- 重构component
+- 基于v4版本的ant-design和complex-component构建模板